feat: 플레이리스트 상세 페이지 디자인 복구 및 리스트 분기 처리 최적화 (#75)
