+++ src/app/playlist/[id]/PlaylistDetail.module.css
@@ -20,6 +20,10 @@
   padding: 4rem 2rem;
 }
 
+.wrapper {
+  width: 100%;
+}
+
 .hero {
   display: flex;
   align-items: flex-end;
